// src/models/chamados.ts
export interface Chamado {
    id?:                any;
    dataAbertura?:   string;
    dataFechamento?: string;
    prioridade:      string;
    status:          string;
    titulo:          string;
    observacao:     string;
    tecnico:            any;
    cliente:            any;
    nomeCliente:     string;
    nomeTecnico:     string;
}